code: use osenv.home()
PR #77 improves the code to not hardcode the home directory to $HOME (which
is only a convention on Unix-like systems). This further improves that
by using `osenv` / `os-homedir`, which have more complete and robust
logic for finding a user's home directory on network shares and other
similar places.

Credit: @othiym23
PR-URL: https://github.com/npm/nopt/pull/77